+++ src/hooks/index.ts
@@ -3,6 +3,7 @@ export * from './use-me';
 export * from './use-device-registration';
 export * from './use-complete-onboarding';
 export * from './use-update-me';
+export * from './use-reduce-motion';
 export * from './use-refresh-control';
 export * from './auth/use-email-sign-in';
 export * from './auth/use-email-sign-up';